Adición del módulo cuestionarios, creación de los controladores (Creación de cuestionario, obtener todos los cuestionarios de un curso, obtener un cuestionario para actualizar, actualizar cuestionario, eliminar cuestionario), adición en 'professor.routes.js' los endpoints de backend y 'professor.js' los endpoints del frontend lo relaciado a cuestionarios, adición de 'questionnaire.schema.js' que se encarga de validar todo el cuestionario (Validaciones de fecha, preguntas vacías, respuestas vacías, preguntas sin respuestas correctas, etc). Implementación de todas las vistas relacionadas al módulo cuestionarios dentro de 'ProfessorPage.jsx'. Por último, se implementan los modelos relacionados al cuestionario 'Questionnaires, Questions, Answers, QuestionsAnswers'.

+++ src/middlewares/validate.js
@@ -1,12 +1,14 @@
 import fs from 'fs'
 
 export const validate = (schema) => async (req, res, next) => {
-    try {
-      console.log(req.body)
-      await schema.parse(req.body)
-      next()
-    } catch (error) {
+  try {
+    console.log(req.body)
+    await schema.parse(req.body)
+    next()
+  } catch (error) {
+    if (req.file) {
       fs.unlinkSync(req.file.path)
-      return res.status(400).json(error.errors.map((error) => error.message))
     }
+    return res.status(400).json(error.errors.map((error) => error.message))
   }
+}
